+++ iris-fsw-softconsole/include/drivers/protocol/can.h
@@ -23,10 +23,10 @@
 //-------------------------------------------------------------------------------------------------------------------------------------------------------------
 // INCLUDES
 //-------------------------------------------------------------------------------------------------------------------------------------------------------------
+#include <FreeRTOS/FreeRTOS-Source/include/FreeRTOS.h>
+#include <FreeRTOS/FreeRTOS-Source/include/queue.h>
 #include "drivers/mss_can/mss_can.h"
 
-#include "FreeRTOS.h"
-#include "queue.h"
 
 //-------------------------------------------------------------------------------------------------------------------------------------------------------------
 // DEFINITIONS AND MACROS
