+++ iris-fsw-softconsole/include/drivers/protocol/spi.h
@@ -27,10 +27,10 @@
 //-------------------------------------------------------------------------------------------------------------------------------------------------------------
 // INCLUDES
 //-------------------------------------------------------------------------------------------------------------------------------------------------------------
+#include <FreeRTOS/FreeRTOS-Source/include/FreeRTOS.h>
+#include <FreeRTOS/FreeRTOS-Source/include/semphr.h>
 #include "drivers/CoreSPI/core_spi.h"   // Contains the interface for the CoreSPI drivers.
 #include "drivers/mss_gpio/mss_gpio.h"
-#include "FreeRTOS.h"
-#include "semphr.h"
 
 //-------------------------------------------------------------------------------------------------------------------------------------------------------------
 // DEFINES
