+++ iris-fsw-softconsole/src/request_code.c
@@ -14,12 +14,11 @@
 //-------------------------------------------------------------------------------------------------------------------------------------------------------------
 // INCLUDES
 //-------------------------------------------------------------------------------------------------------------------------------------------------------------
+#include <FreeRTOS/FreeRTOS-Source/include/FreeRTOS.h>
+#include <FreeRTOS/FreeRTOS-Source/include/task.h>
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
-#include <FreeRTOS.h>
-#include <task.h>
-
 #include "request_code.h"
 
 //-------------------------------------------------------------------------------------------------------------------------------------------------------------
