+++ iris-fsw-softconsole/src/tasks/priority_queue.c
@@ -15,12 +15,11 @@
 //-------------------------------------------------------------------------------------------------------------------------------------------------------------
 // INCLUDES
 //-------------------------------------------------------------------------------------------------------------------------------------------------------------
+#include <FreeRTOS/FreeRTOS-Source/include/FreeRTOS.h>
+#include <FreeRTOS/FreeRTOS-Source/include/task.h>
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
-#include <FreeRTOS.h>
-#include <task.h>
-
 #include "tasks/priority_queue.h"
 
 //-------------------------------------------------------------------------------------------------------------------------------------------------------------
