+++ iris-fsw-softconsole/tests/adcs_tests.c
@@ -12,10 +12,9 @@
 // - Created.
 //-------------------------------------------------------------------------------------------------
 
+#include <FreeRTOS/FreeRTOS-Source/include/FreeRTOS.h>
+#include <FreeRTOS/FreeRTOS-Source/include/task.h>
 #include "tests.h"
-
-#include "FreeRTOS.h"
-#include "task.h"
 
 #include "drivers/device/adcs_driver.h"
