+++ iris-fsw-softconsole/tests/can_tests.c
@@ -12,10 +12,10 @@
 // - Created.
 //-------------------------------------------------------------------------------------------------
 
+#include <FreeRTOS/FreeRTOS-Source/include/FreeRTOS.h>
+#include <FreeRTOS/FreeRTOS-Source/include/task.h>
 #include "tests.h"
 
-#include "FreeRTOS.h"
-#include "task.h"
 #include "drivers/protocol/can.h"
